fix:transaction product

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UpdateUserRequest;
 use App\Models\Cart;
 use App\Models\Market;
+use App\Models\Transaction;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -103,7 +104,13 @@ class UserController extends Controller
 
         $user = User::findOrFail($id);
         if (Market::where('user_id', $user->id)->exists()) {
-            return redirect('/users')->with('error', 'User sudah terdaftar di sebagai admin toko!');
+            return redirect('/users')->with('error', 'User telah terdaftar di sebagai admin toko!');
+        }
+        if (Transaction::where('user_id', $user->id)->exists()) {
+            return redirect('/users')->with('error', 'User telah melakukan transaksi!');
+        }
+        if (Cart::where('user_id', $user->id)->exists()) {
+            return redirect('/users')->with('error', 'User telah menambahkan produk ke keranjang!');
         }
         $user->delete();
         return redirect('/users')->with('error', 'User berhasil dihapus!');
